rename EventDetails and DeadlineDetails
