+++ src-tauri/src/event_structures/mod.rs
@@ -1,3 +1,3 @@
-pub mod deadline_details;
-pub mod event_details;
+pub mod calendar_deadline;
+pub mod calendar_event;
 pub mod calendar_entry;
